Apply eslint onto existing codebase
