+++ frontend/babel.config.js
@@ -1,4 +1,5 @@
-module.exports = function(api) {
+/* eslint-disable func-names */
+module.exports = function (api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
