+++ frontend/package.json
@@ -6,7 +6,7 @@
     "ios": "expo start --ios",
     "web": "expo start --web",
     "eject": "expo eject",
-    "test": "standard"
+    "test": "npx eslint '**/*.js' '**/*.jsx'"
   },
   "dependencies": {
     "@expo/vector-icons": "^12.0.2",
