+++ frontend/style.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-export const styles = StyleSheet.create({
+const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: '#fff',
@@ -47,5 +47,7 @@ export const styles = StyleSheet.create({
     padding: 20,
     alignItems: 'stretch',
     justifyContent: 'center',
-  }
+  },
 });
+
+export default styles;
